requirements.txt
flask==3.0.3
google-cloud-bigquery==3.25.0
google-api-core==2.19.1
prometheus_client==0.20.0
gunicorn==23.0.0
